PDF_Slide_Extractor/requirements.txt
PyMuPDF==1.23.26
Pillow==10.0.1
pathlib2==2.3.7
